+++ arrow-fx-coroutines-stream/src/main/kotlin/arrow/fx/coroutines/stream/predef.kt
@@ -1,5 +1,7 @@
 package arrow.fx.coroutines.stream
 
+import arrow.fx.coroutines.IQueue
+
 internal fun checkBounds(arraySize: Int, offset: Int, length: Int) {
   require(offset in 0..arraySize)
   require(length in 0..arraySize)
@@ -36,7 +38,45 @@ internal fun <A, B> Iterator<A>.flatMap(f: (A) -> Iterator<B>): Iterator<B> =
       (if (hasNext()) cur else empty).next()
   }
 
+internal const val ArrowExceptionMessage =
+  "Arrow-kt internal error. Please let us know and create a ticket at https://github.com/arrow-kt/arrow/issues/new/choose"
+
+internal class ArrowInternalException(override val message: String = ArrowExceptionMessage) : RuntimeException(message)
+
+internal fun <A> IQueue<A>.tail(): IQueue<A> =
+  tailOrNull() ?: throw ArrowInternalException("$ArrowExceptionMessage\nTail on empty queue.")
+
 /** Represents a unique identifier using object equality. */
 internal class Token {
   override fun toString(): String = "Token(${Integer.toHexString(hashCode())})"
 }
+
+infix fun <A> A.prependTo(fa: Iterable<A>): List<A> =
+  listOf(this) + fa
+
+internal fun <A> Iterable<A>.deleteFirst(f: (A) -> Boolean): Pair<A, List<A>>? {
+  tailrec fun go(rem: Iterable<A>, acc: List<A>): Pair<A, List<A>>? =
+    when {
+      rem.isEmpty() -> null
+      else -> {
+        val a = rem.first()
+        val tail = rem.drop(1)
+        if (!f(a)) go(tail, acc + a)
+        else Pair(a, acc + tail)
+      }
+    }
+
+  return go(this, emptyList())
+}
+
+internal fun <A> Iterable<A>.uncons(): Pair<A, List<A>>? =
+  firstOrNull()?.let { Pair(it, drop(1)) }
+
+internal fun Iterable<*>.isEmpty(): Boolean =
+  size() == 0
+
+internal fun Iterable<*>.size(): Int =
+  when (this) {
+    is Collection -> size
+    else -> fold(0) { acc, _ -> acc + 1 }
+  }
