+++ arrow-fx-coroutines/src/main/kotlin/arrow/fx/coroutines/IQueue.kt
@@ -31,9 +31,6 @@ data class IQueue<A> internal constructor(
   internal fun first(): A =
     firstOrNull() ?: throw NoSuchElementException("first on empty queue")
 
-  internal fun tail(): IQueue<A> =
-    tailOrNull() ?: throw NoSuchElementException("tail on empty queue")
-
   fun firstOrNull(): A? =
     when {
       listOut.isNotEmpty() -> listOut.first()
