+++ gradle.properties
@@ -8,7 +8,6 @@ PUBLICATION=https://raw.githubusercontent.com/arrow-kt/arrow/master/gradle/publi
 # Gradle options
 org.gradle.jvmargs=-Xmx4g
 org.gradle.parallel=true
-org.gradle.workers.max=2
 # Kotlin configuration
 kotlin.incremental=true
 # Kotlin Test configuration
